[MODIFY]Member->User로 바꾸기
